+++ CHANGELOG.md
@@ -7,18 +7,21 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - Configurable Agent API calls timeout.
+- New `AGENT_INFO_NEW` event.
 - New `NODE_NEW` event.
 - New `SHARD_ALLOCATION_NEW` event.
 
 ### Changed
 - **BREAKING**: Flatten encoded events structure (`data`, `event`, `timeout` are all root attributes).
-- **BREAKING**: Update models.
 - **BREAKING**: Rename `*Recover` events.
 - **BREAKING**: Rename `DatastoreDown` to `NodeDown`.
 - **BREAKING**: Rename `DatastoreRecovered` to `NodeUp`.
 - **BREAKING**: Rename `EventData` to `EventPayload`.
 - **BREAKING**: Replace `AgentStillDown` with `AgentDown`.
 - **BREAKING**: Replace `DatastoreStillDown` with `NodeDown`.
+- **BREAKING**: Rework `AGENT_NEW` to include only cluster and host.
+- **BREAKING**: Update models.
+- Emit agent status change after emitting `AGENT_NEW` events.
 - Move logging code to common crate.
 
 ### Fixed